comps/imagegen/requirements.txt
docarray[full]
fastapi
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-sdk
sentencepiece
shortuuid
torch
tritonclient[http]
numpy
pillow
